requirements.txt
flask>=3.0
requests>=2.31
anthropic>=0.30
gunicorn>=21.0
supabase>=2.0
python-dotenv>=1.0
